requirements.txt
colorama==0.4.6
PyYAML==6.0.2
tensorflow==2.19.0
numpy==2.1.3
pylsl==1.17.6
pytest==8.3.5
pyserial==3.5
pyqtgraph==0.13.7
PySide6==6.9.0
medusa-kernel==1.4.0